requirements.txt
numpy==1.21.5
pandas==1.3.5
python-dotenv==0.21.0
scikit_bio==0.5.7
scikit_learn==1.1.3
scikit_survival==0.17.2
scipy==1.7.3
synapseclient==2.7.0
